mastodon_harvesters/hashtag_harvester/requirements.txt
beautifulsoup4==4.11.1
CouchDB==1.2
PyYAML==6.0
PyYAML==6.0
spacy==3.5.3
Mastodon.py==1.8.1
requests==2.30.0